refactor(PUPIL-1753): dedupe results toggle aria-label wrapper
Extract shared LessonReviewQuizWithAriaLabel and use display: contents
to preserve layout while labelling the internal results toggle.
