+++ src/components/PupilComponents/Views/PupilLessonReview/LessonReviewQuizWithAriaLabel/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./LessonReviewQuizWithAriaLabel";
+export type { LessonReviewQuizWithAriaLabelProps } from "./LessonReviewQuizWithAriaLabel";
